[feature] update how output feature is represented
Originally each line of the feature output file correponds to the
entire graph's node features. This leads to lots of duplication
when only a single node's feature changes.

This PR changes each line to only output a single feature that was
changed. It's up to users to decide how to parse the features.

Also, updated the feature template to output timestamp first
example output:
```
0,A,0,1,10,2,1,2,1,2
0,C,1,3,0,3,2,1,2,2
0,E,1,0,0,3,1,1,2,1
2,B,1,3,0,3,2,1,2,2
2,D,1,0,0,3,1,1,2,1
10,A,0,1,10,2,1,3,1,3
10,C,1,3,0,3,2,1,3,3
12,B,1,3,0,3,2,1,3,3
12,E,1,0,0,3,1,1,3,1
14,D,1,0,0,3,1,1,3,1
20,A,0,1,10,2,1,4,1,4
21,B,1,3,0,3,2,1,4,4
21,C,1,3,0,3,2,1,4,4
21,D,1,0,0,3,1,1,4,1
21,E,1,0,0,3,1,1,4,1
30,A,0,1,10,2,1,5,1,5
31,C,1,3,0,3,2,1,5,5
32,B,1,3,0,3,2,1,5,5
32,E,1,0,0,3,1,1,5,1
34,D,1,0,0,3,1,1,5,1
40,A,0,1,10,2,1,6,1,6
40,C,1,3,0,3,2,1,6,6
41,E,1,0,0,3,1,1,6,1
42,B,1,3,0,3,2,1,6,6
44,D,1,0,0,3,1,1,6,1
```